Add the preference path: supervise the routing decision directly
Priority 0 needs an empirical answer, so this builds the alternative it
compares against.

New source `router.data.sources.preference` turns LMArena battles into a
binary `routing_label` (strong_needed / weak_sufficient), following the
RouteLLM framing. Model tiers are *derived from observed win rates*, not
hand-assigned -- labelling 50+ models by reputation would bake the author's
priors into the ground truth. Only mixed-tier battles are kept; a
strong-vs-strong pairing says nothing about whether the cheap path would
have sufficed. Ties count as weak_sufficient on purpose: if a cheap model
draws with an expensive one, routing cheap is the correct decision.

Yields 24,903 / 5,337 / 5,337 rows, balanced 51/49 -- 4x the domain set, on
real user prompts rather than academic MCQs.

Also adds `min_tier_gap`, since the label mixes battles with very different
strong/weak gaps (widest third 57% strong_needed vs narrowest third 42%)
and that gap is a property of the pairing the classifier never sees. Note
the filter did *not* help in practice (AUC 0.549 -> 0.549), so pairing noise
is not what limits this task.

Supporting changes:
- `routing_label` added to the canonical schema, so both label spaces share
  one row type and the model/metric/experiment layers are untouched. That
  the harness needed no changes is the claim of label-agnosticism holding up.
- ROC-AUC and PR-AUC for binary heads: the router thresholds a score, so
  ranking quality matters more than accuracy at an arbitrary 0.5 cut.
- `compare_paths` scores direct and indirect routing against the same ground
  truth on the same rows, with a majority-class baseline always reported.

Fixes a real bug found by experiment 12: DeBERTa-v3 loads in fp16 and
crashed the class-weighted loss with "expected scalar type Half but found
Float". The loss is now always computed in fp32. Regression-tested.

+++ src/router/models/finetune.py
@@ -113,7 +113,11 @@ class FineTunedTransformer(DomainClassifier):
             for batch in loader:
                 batch = {k: v.to(self.device) for k, v in batch.items()}
                 targets = batch.pop("labels")
-                logits = self.model(**batch).logits
+                # Always compute the loss in fp32. Some checkpoints (DeBERTa-v3)
+                # load in half precision, and a Half/Float mismatch against the
+                # class-weight tensor fails outright; fp32 loss is also the
+                # numerically safer default.
+                logits = self.model(**batch).logits.float()
                 loss = criterion(logits, targets)
                 loss.backward()
                 torch.nn.utils.clip_grad_norm_(self.model.parameters(), 1.0)
